requirements.txt
pygit2
psycopg2
tqdm
pyyaml
